refactor liveblog logic
